I wrote a new GrabString Function, UARTGrabString, that receives a string from USART0. A driver program is in uprintf.c
