+++ AVR_LIBRARIES/uart.c
@@ -6,10 +6,6 @@
 //////////////////////////////////////////////////////
 
 
-
-#include <avr/io.h>
-#include <avr/pgmspace.h>
-
 #include "uart.h"
 
 void uartInit(uint32_t baudrate)
@@ -87,3 +83,32 @@ void UARTGrabString (char* stringBuffer, int maxStringSize)
     stringBuffer[i] = NULL_CHAR;
     // END STRING RECEIVING LOGIC
 }
+
+
+///* CREATED BY SARGIS S YONAN - 12 OCT. 2015 */
+///* A FUNCTION THAT MIMICS <stdio.h>'s PRINTF FUNCTION */
+///* FOR THE AVR ATMEGA 2560 MICROCONTROLLER */
+
+void uprintf (char* input_string, ...)
+{
+    va_list valist;
+    char* newString;
+    int stringLength = 0;
+    
+    va_start(valist, input_string);
+    
+    for (stringLength = 0; input_string[stringLength] != '\0'; ++stringLength) { }
+    
+    newString = (char*)malloc(stringLength *  STRING_MULTIPLIER);
+    vsprintf(newString, input_string, valist);
+    
+    ////////////////////////////
+    // WRITING TO UART STREAM //
+    ////////////////////////////
+    OUTPUT_FUNCTION(newString);
+    free(newString);
+    va_end(valist);
+}
+
+
+
