+++ AVR_LIBRARIES/uart.h
@@ -1,10 +1,17 @@
 #define F_CPU 16000000UL
 #include <string.h>
+#include <avr/io.h>
+#include <avr/pgmspace.h>
+#include <stdarg.h>
+#include <stdio.h>
+#include <stdlib.h>
 
 #define NULL_CHAR '\0'
 #define NEWLINE_CHAR '\n'
 #define BACKSPACE_CHAR 0x08
 
+#define STRING_MULTIPLIER 4
+#define OUTPUT_FUNCTION uartPutString
 
 #ifndef _UART_H_
 #define _UART_H_
@@ -25,6 +32,8 @@ void uartPutString_P(const char*);
 //////////////////////////////////////////////////////
 
 void UARTGrabString (char* stringBuffer, int maxStringSize);
+
+void uprintf (char* input_string, ...);
 
 
 #endif /* _UART_H_ */
